requirements.txt
python-sane
makemkv
